+++ src/UnhandledEnumException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Zlikavac32\Enum;
+namespace Zlikavac32\ZEnum;
 
 use LogicException;
 
@@ -11,7 +11,7 @@ class UnhandledEnumException extends LogicException
 
     private string $enumFqn;
 
-    public function __construct(public Enum $enum)
+    public function __construct(public ZEnum $enum)
     {
         $fqn = get_parent_class($enum);
 
@@ -20,7 +20,7 @@ class UnhandledEnumException extends LogicException
         $this->enumFqn = $fqn;
     }
 
-    public function enum(): Enum
+    public function enum(): ZEnum
     {
         return $this->enum;
     }