+++ tests/Fixtures/DuplicateNameEnum.php
@@ -2,15 +2,15 @@
 
 declare(strict_types=1);
 
-namespace Zlikavac32\Enum\Tests\Fixtures;
+namespace Zlikavac32\ZEnum\Tests\Fixtures;
 
-use Zlikavac32\Enum\Enum;
+use Zlikavac32\ZEnum\ZEnum;
 
 /**
  * @method static DuplicateNameEnum ENUM_A
  * @method static DuplicateNameEnum ENUM_A
  */
-abstract class DuplicateNameEnum extends Enum
+abstract class DuplicateNameEnum extends ZEnum
 {
 
 }