+++ tests/Fixtures/EnumThatExtendsValidObjectsEnum.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Zlikavac32\Enum\Tests\Fixtures;
+namespace Zlikavac32\ZEnum\Tests\Fixtures;
 
 /**
  * @method static EnumThatExtendsValidObjectsEnum ENUM_A