+++ tests/Fixtures/InvalidObjectAliasEnumerationObjectsEnum.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Zlikavac32\Enum\Tests\Fixtures;
+namespace Zlikavac32\ZEnum\Tests\Fixtures;
 
-use Zlikavac32\Enum\Enum;
+use Zlikavac32\ZEnum\ZEnum;
 
 /**
  * @method static InvalidObjectAliasEnumerationObjectsEnum ENUM
  */
-abstract class InvalidObjectAliasEnumerationObjectsEnum extends Enum
+abstract class InvalidObjectAliasEnumerationObjectsEnum extends ZEnum
 {
     protected static function enumerate(): array
     {